# Add option to decode CBOR bignum to `interface{}` as `*big.Int`