Closes #444 

This PR adds decoding option `BigIntDecMode` which specifies how to decode CBOR bignum to Go `interface{}`.

- `BigIntDecodeValue` (default) makes CBOR bignum decode to `big.Int`.
- `BigIntDecodePointer` makes CBOR bignum decode to `*big.Int`.

Default behavior is unchanged to maintain backward compatibility.  However, using the new option is recommended as it improves speed, reduces memory use, and is more ergonomic.

### Benchmark Comparison

```
                  │ bench_decode_to_bigint.txt │     bench_decode_to_pbigint.txt      │
                  │           sec/op           │    sec/op     vs base                │
DecodeToBigInt-12                  238.9n ± 8%   196.2n ± 10%  -17.87% (p=0.000 n=20)

                  │ bench_decode_to_bigint.txt │    bench_decode_to_pbigint.txt     │
                  │            B/op            │    B/op     vs base                │
DecodeToBigInt-12                   144.0 ± 0%   112.0 ± 0%  -22.22% (p=0.000 n=20)

                  │ bench_decode_to_bigint.txt │    bench_decode_to_pbigint.txt     │
                  │         allocs/op          │ allocs/op   vs base                │
DecodeToBigInt-12                   5.000 ± 0%   4.000 ± 0%  -20.00% (p=0.000 n=20)
```

### Special Thanks

Many thanks to @extemporalgenome for opening issue #444!  :+1: