@fxamacker 2023-12-28
@x448 Thanks!  I added benchmark comparisons to the PR text.